# perf: use allocUnsafe in node