@marcus-pousette 2023-04-26
I realized, the benchmark above is invalid. There are actually no returns to be gained from allocUnsafe since the header is just 12 bytes. In fact, protobufjs has a special case for small arrays where they use Uint8Array instead of Buffer because it is faster when size < 40 (isch). 

The issue with the benchmark above is that it provided results when the output was wrong, hence the egregious difference. (frame[0] = 0 was never set because there was an assumption on this when using only Uint8array). Doing this correctly, yields no improvements whatsoever.  

I am going to close this PR and look for other improvements.